All parametric tests are green (still 17 failures of other tests)

+++ src/main/java/org/biokoframework/systema/entity/dummyComplex/DummyComplexDomainEntity.java
@@ -38,7 +38,7 @@ public class DummyComplexDomainEntity extends DomainEntity {
 	public static final String A_STRING_MANDATORY_FIELD = "aStringMandatoryField";
 	@Field(mandatory = false)
 	public static final String A_STRING_OPTIONAL_FIELD = "aStringOptionalField";
-	@Field(type = Integer.class, mandatory = false)
+	@Field(type = Long.class, mandatory = false)
 	public static final String AN_INTEGER_OPTIONAL_FIELD = "anIntegerOptionalField";
 	@Field
 	public static final String A_STRING_FIELD_MANDATORY_ALSO_IN_GET = "aStringFieldMandatoryAlsoInGet";
